refactor: move envets to the relevent packages
