+++ packages/telemetry/package.json
@@ -19,8 +19,7 @@
     "is-wsl": "2.2.0",
     "ci-info": "3.6.2",
     "async-retry": "1.3.3",
-    "node-fetch": "2.6.7",
-    "find-up": "4.1.0"
+    "node-fetch": "2.6.7"
   },
   "devDependencies": {
     "@types/async-retry": "1.3.0",
